+++ pcigale/session/configuration.py
@@ -177,9 +177,9 @@ class Configuration:
 
         modules = self.config["sed_modules"]
         if "m2005" in modules:
-            if "nebular" in modules or "xray" in modules:
+            if "xray" in modules:
                 raise Exception("The m2005 module is not compatible with the "
-                                "nebular and xray modules.")
+                                "xray module.")
         # Getting the list of the filters available in pcigale database
         with Database("filters") as db:
             filter_list = db.parameters["name"]